Add files via upload
The objective of this project is to build a tic-tac-toe game (also known as X-O game) so anyone can play without depleting paper. An individual can play against the computer. The player who succeeds in placing their marks in a diagonal, horizontal, or vertical row is the winner. Developed using C++ programming language.